Changed: Moved all Split/Unsplit Blocks in BC3 Module to Shared Prelude
