+++ projects/dxt-lossless-transform-bc3/src/split_blocks/mod.rs
@@ -136,52 +136,3 @@ pub unsafe fn unsplit_blocks(input_ptr: *const u8, output_ptr: *mut u8, len: usi
     debug_assert!(len % 16 == 0);
     unsplit::unsplit_blocks(input_ptr, output_ptr, len);
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::split_blocks::split::tests::{
-        generate_bc3_test_data, transform_with_reference_implementation,
-    };
-    use crate::testutils::allocate_align_64;
-    use rstest::rstest;
-
-    #[rstest]
-    #[case::min_size(1)] // 8 bytes - minimum size
-    #[case::min_size(2)] // 16 bytes - SSE Register
-    #[case::min_size(4)] // 32 bytes - AVX Register
-    #[case::min_size(8)] // 64 bytes - SSE Unrolled Operation
-    #[case::min_size(16)] // 128 bytes - AVX Unrolled Operation
-    #[case::min_size(32)] // 256 bytes - Multiple Unrolled Operations
-    fn test_transform_untransform(#[case] num_blocks: usize) {
-        let input = generate_bc3_test_data(num_blocks);
-        let mut transformed = allocate_align_64(input.len());
-        let mut reconstructed = allocate_align_64(input.len());
-        let mut reference = allocate_align_64(input.len());
-
-        // Generate reference output
-        transform_with_reference_implementation(input.as_slice(), reference.as_mut_slice());
-
-        unsafe {
-            // Test transform
-            split_blocks(input.as_ptr(), transformed.as_mut_ptr(), input.len());
-            assert_eq!(
-                transformed.as_slice(),
-                reference.as_slice(),
-                "transform_bc3 produced different results than reference for {num_blocks} blocks"
-            );
-
-            // Test untransform
-            unsplit_blocks(
-                transformed.as_ptr(),
-                reconstructed.as_mut_ptr(),
-                transformed.len(),
-            );
-            assert_eq!(
-                reconstructed.as_slice(),
-                input.as_slice(),
-                "untransform_bc3 failed to reconstruct original data for {num_blocks} blocks"
-            );
-        }
-    }
-}
