+++ projects/dxt-lossless-transform-bc3/src/split_blocks/split/avx2.rs
@@ -211,125 +211,15 @@ unsafe fn write_alpha_bits_u64(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::split_blocks::split::tests::{
-        assert_implementation_matches_reference, generate_bc3_test_data,
-        transform_with_reference_implementation,
-    };
-    use crate::testutils::allocate_align_64;
-    use rstest::rstest;
-
-    #[rstest]
-    fn test_avx2_aligned() {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
-            return;
-        }
-
-        for num_blocks in 1..=512 {
-            let input = generate_bc3_test_data(num_blocks);
-            let mut output_expected = vec![0u8; input.len()];
-            transform_with_reference_implementation(input.as_slice(), &mut output_expected);
-
-            let mut output_test = allocate_align_64(input.len());
-            output_test.as_mut_slice().fill(0);
-            unsafe {
-                u32_avx2(input.as_ptr(), output_test.as_mut_ptr(), input.len());
-            }
-
-            assert_implementation_matches_reference(
-                output_expected.as_slice(),
-                output_test.as_slice(),
-                "avx2",
-                num_blocks,
-            );
-        }
-    }
+    use crate::test_prelude::*;
 
     #[rstest]
     fn test_avx2_unaligned() {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
+        if !has_avx2() {
             return;
         }
 
-        for num_blocks in 1..=512 {
-            let input = generate_bc3_test_data(num_blocks);
-            let mut output_expected = vec![0u8; input.len() + 1];
-            transform_with_reference_implementation(input.as_slice(), &mut output_expected[1..]);
-
-            let mut output_test = vec![0u8; input.len() + 1];
-            output_test.as_mut_slice().fill(0);
-            unsafe {
-                u32_avx2(input.as_ptr(), output_test.as_mut_ptr().add(1), input.len());
-            }
-
-            assert_implementation_matches_reference(
-                &output_expected[1..],
-                &output_test[1..],
-                "avx2",
-                num_blocks,
-            );
-        }
-    }
-
-    #[rstest]
-    fn avx2_separate_pointers_matches_contiguous() {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
-            return;
-        }
-
-        for num_blocks in 1..=256 {
-            let input = generate_bc3_test_data(num_blocks);
-            let len = input.len();
-
-            // Test with the contiguous buffer method
-            let mut output_contiguous = allocate_align_64(len);
-
-            // Test with separate pointers
-            let mut alpha_bytes = allocate_align_64(len / 8); // 2 bytes per block
-            let mut alpha_bits = allocate_align_64(len * 3 / 8); // 6 bytes per block
-            let mut colors = allocate_align_64(len / 4); // 4 bytes per block
-            let mut indices = allocate_align_64(len / 4); // 4 bytes per block
-
-            unsafe {
-                // Reference: contiguous buffer
-                u32_avx2(input.as_ptr(), output_contiguous.as_mut_ptr(), len);
-
-                // Test: separate pointers
-                u32_avx2_with_separate_pointers(
-                    input.as_ptr(),
-                    alpha_bytes.as_mut_ptr() as *mut u16,
-                    alpha_bits.as_mut_ptr(),
-                    colors.as_mut_ptr() as *mut u32,
-                    indices.as_mut_ptr() as *mut u32,
-                    (alpha_bytes.as_mut_ptr() as *mut u16).add(len / 16),
-                );
-            }
-
-            // Verify that separate pointer results match contiguous buffer layout
-            let expected_alpha_bytes = &output_contiguous.as_slice()[0..len / 8];
-            let expected_alpha_bits = &output_contiguous.as_slice()[len / 8..len / 2];
-            let expected_colors = &output_contiguous.as_slice()[len / 2..len * 3 / 4];
-            let expected_indices = &output_contiguous.as_slice()[len * 3 / 4..];
-
-            assert_eq!(
-                alpha_bytes.as_slice(),
-                expected_alpha_bytes,
-                "AVX2 Alpha bytes section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                alpha_bits.as_slice(),
-                expected_alpha_bits,
-                "AVX2 Alpha bits section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                colors.as_slice(),
-                expected_colors,
-                "AVX2 Color section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                indices.as_slice(),
-                expected_indices,
-                "AVX2 Index section mismatch for {num_blocks} blocks"
-            );
-        }
+        // For AVX2: processes 128 bytes (8 blocks) per iteration, so max_blocks = 128 bytes × 2 ÷ 16 = 16
+        run_standard_transform_unaligned_test(u32_avx2, 16, "avx2");
     }
 }
