+++ projects/dxt-lossless-transform-bc3/src/split_blocks/unsplit/avx512.rs
@@ -808,94 +808,21 @@ pub unsafe fn avx512_detransform_separate_components_32_vl(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::split_blocks::split::tests::generate_bc3_test_data;
-    use crate::split_blocks::split::u32;
-    use crate::split_blocks::unsplit::tests::assert_implementation_matches_reference;
-    use crate::testutils::allocate_align_64;
-    use rstest::rstest;
-
-    type DetransformFn = unsafe fn(*const u8, *mut u8, usize);
+    use crate::test_prelude::*;
 
     #[rstest]
-    #[case(avx512_detransform, "avx512")]
-    #[case(avx512_detransform_32_vl, "avx512_32_vl")]
-    #[case(avx512_detransform_32_vbmi, "avx512_32_vbmi")]
-    fn test_avx512_aligned(#[case] detransform_fn: DetransformFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512vbmi()
-            || !dxt_lossless_transform_common::cpu_detect::has_avx512vl()
-        {
+    #[case(avx512_detransform, "avx512", 64)] // main processes 512 bytes (32 blocks), so max_blocks = 512 × 2 ÷ 16 = 64
+    #[case(avx512_detransform_32_vbmi, "avx512_32_vbmi", 8)] // _32 variants process 64 bytes (4 blocks), so max_blocks = 64 × 2 ÷ 16 = 8
+    #[case(avx512_detransform_32_vl, "avx512_32_vl", 8)] // _32 variants process 64 bytes (4 blocks), so max_blocks = 64 × 2 ÷ 16 = 8
+    fn test_avx512_unaligned(
+        #[case] detransform_fn: StandardTransformFn,
+        #[case] impl_name: &str,
+        #[case] max_blocks: usize,
+    ) {
+        if !has_avx512vbmi() || !has_avx512vl() {
             return;
         }
 
-        for num_blocks in 1..=512 {
-            let original = generate_bc3_test_data(num_blocks);
-            let mut transformed = allocate_align_64(original.len());
-            let mut reconstructed = allocate_align_64(original.len());
-
-            unsafe {
-                // Transform using standard implementation
-                u32(original.as_ptr(), transformed.as_mut_ptr(), original.len());
-
-                // Reconstruct using the implementation being tested
-                reconstructed.as_mut_slice().fill(0);
-                detransform_fn(
-                    transformed.as_ptr(),
-                    reconstructed.as_mut_ptr(),
-                    transformed.len(),
-                );
-            }
-
-            assert_implementation_matches_reference(
-                original.as_slice(),
-                reconstructed.as_slice(),
-                &format!("{impl_name} (aligned)"),
-                num_blocks,
-            );
-        }
-    }
-
-    #[rstest]
-    #[case(avx512_detransform, "avx512")]
-    #[case(avx512_detransform_32_vbmi, "avx512_32_vbmi")]
-    #[case(avx512_detransform_32_vl, "avx512_32_vl")]
-    fn test_avx512_unaligned(#[case] detransform_fn: DetransformFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512vbmi()
-            || !dxt_lossless_transform_common::cpu_detect::has_avx512vl()
-        {
-            return;
-        }
-
-        for num_blocks in 1..=512 {
-            let original = generate_bc3_test_data(num_blocks);
-
-            // Transform using standard implementation
-            let mut transformed = vec![0u8; original.len()];
-            unsafe {
-                u32(original.as_ptr(), transformed.as_mut_ptr(), original.len());
-            }
-
-            // Add 1 extra byte at the beginning to create misaligned buffers
-            let mut transformed_unaligned = vec![0u8; transformed.len() + 1];
-            transformed_unaligned[1..].copy_from_slice(&transformed);
-
-            let mut reconstructed = vec![0u8; original.len() + 1];
-
-            unsafe {
-                // Reconstruct using the implementation being tested with unaligned pointers
-                reconstructed.as_mut_slice().fill(0);
-                detransform_fn(
-                    transformed_unaligned.as_ptr().add(1),
-                    reconstructed.as_mut_ptr().add(1),
-                    transformed.len(),
-                );
-            }
-
-            assert_implementation_matches_reference(
-                original.as_slice(),
-                &reconstructed[1..],
-                &format!("{impl_name} (unaligned)"),
-                num_blocks,
-            );
-        }
+        run_standard_untransform_unaligned_test(detransform_fn, max_blocks, impl_name);
     }
 }
